\name{gen_param_mc}
\alias{gen_param_mc}
\title{Generate parameter sets for calculating risk projections with uncertainty}
\description{
Convenience function to generate a list of parameter sets for use in lifetime excess absolute risk functions that take into account uncertainty.
}
\usage{
gen_param_mc(exposure, n_sim = 1000L,
             wt_transfer, lat_t0, lat_eta,
             lat_method = c("ProZES", "RadRAT"),
             lat_fixed = FALSE, wt_transfer_fixed = FALSE,
             ddref_fixed = FALSE, risk_model, risk_model_mort)
}
\arguments{
  \item{exposure}{Parameters associated with the exposure event. See \code{\link{gen_exposure}} for convenience function to define. A list with one component for each exposure event. Each exposure event component a list with the following components:
      \itemize{
        \item \code{sex}: Character, \code{"f"} or \code{"m"} - must be constant over exposure events.
        \item \code{agex}: Age at exposure.
        \item \code{dose_distr}: Distribution to use for generating a dose for each exposure event. See \code{\link{sim_exposure}}.
        \item \code{dose_param}: Parameter vector to use for generating \code{n} dose values in Gy or Sv, depending on \code{dose_distr}. See \code{\link{sim_exposure}}.
        \item \code{dose_rate}: CAVE: Currently unused. Supported: \code{"acute"}, \code{"chronic"}.
        \item \code{ddref} Dose and dose rate effectiveness factor.
        \item \code{cancer_site}: Character, one or multiple names of relevant cancer sites, corresponding to exposed organs. See \code{\link{get_lear_indiv}}.
        \item \code{ddref}: Dose- and Dose-Rate Effectiveness Factor with a default of 1.
        \item \code{dose_rate}: Character, \code{"acute"} or \code{"chronic"}.
      }
    }
  \item{n_sim}{Numer of simulated parameter settings for Monte Carlo approach to estimate uncertainty interval. See Details.}
  \item{wt_transfer}{A named list with one component per cancer site. Each component a named 2-vector giving the ERR / EAR weights used for risk transfer. This vector must have names \code{"ERR"} and \code{"EAR"}, and weights must be within [0, 1] as well as sum to 1.}
  \item{lat_t0}{A named list with one component per cancer site. Each component the central value For the S-shaped latency function.}
  \item{lat_eta}{A named list with one component per cancer site. Each component the width of the transition period from 0 to full risk for the S-shaped latency function. Used for argument \code{lat_method="ProZES"}.}
  \item{lat_fixed}{Set latency parameters \code{lat_t0} and \code{lat_eta} to a fixed value without uncertainty?}
  \item{wt_transfer_fixed}{Set \code{wt_transfer} weights to a fixed value without uncertainty?}
  \item{ddref_fixed}{Set \code{ddref} Dose- and Dose-Rate Effectiveness Factor to a fixed value without uncertainty?}
  \item{risk_model}{A named list with one component for each cancer site. Definition of ERR and EAR risk models. Cancer sites must correspond to \code{l_param[["cancer_site"]]} and components of \code{base_cancer}. For an example of a risk model for a single cancer site, see \code{\link{rm_solid_incid_walsh2021}}.}
  \item{risk_model_mort}{Only necessary for \code{metric} \code{"REID"} / \code{"REIC"}, \code{"ELR"}. A named list like \code{risk_model} defining ERR and EAR risk models for cancer mortality. For an example, see \code{\link{rm_solid_mort_sumray}}.}
}
\details{
TODO: Describe implementation of Monte Carlo methods for generating sets of parameter values. Specifically, see \code{\link{sim_exposure}} for dose values.
}
\value{
A list with one component for each simulated parameter set. Each such component is a list that can be passed as the first argument to \code{\link{get_lear_indiv}}:
    \itemize{
      \item \code{exposure}: A list with one component for each exposure event as generated by \code{\link{sim_exposure}}. Each exposure event component a list with the following components:
      \itemize{
        \item \code{sex}: Character, \code{"f"} or \code{"m"} - must be constant over exposure events.
        \item \code{agex_timing}: Age at exposure or timing of exposure events (see \code{\link{get_lear_pop}}).
        \item \code{dose}: In Gy or Sv.
        \item \code{ddref}: Dose- and Dose-Rate Effectiveness Factor with a default of 1.
        \item \code{dose_rate}: Character, \code{"acute"} or \code{"chronic"}.
        \item \code{cancer_site}: Character, one or multiple names of relevant cancer sites, corresponding to exposed organs. See \code{\link{get_lear_indiv}}.
      }
      \item \code{cancer_site}: A list with one component for each relevant cancer site, corresponding to exposed organs.
      \itemize{
        \item \code{cancer_site}: Character. The name of the cancer site. See \code{\link{get_lear_indiv}}.
        \item \code{param_err}: Parameter vector for the ERR risk model defined in \code{risk_model} for this cancer site.
        \item \code{param_ear}: Parameter vector for the EAR risk model defined in \code{risk_model} for this cancer site.
        \item \code{param_err_mort}: Only necessary for \code{metric} \code{"REID"} / \code{"REIC"}, \code{"ELR"}. Parameter vector for the ERR mortality risk model defined in \code{risk_model_mort} for this cancer site.
        \item \code{param_ear_mort}: Only necessary for \code{metric} \code{"REID"} / \code{"REIC"}, \code{"ELR"}. Parameter vector for the EAR mortality risk model defined in \code{risk_model_mort} for this cancer site.
        \item \code{wt_transfer}: A named 2-vector giving the ERR / EAR weights used for risk transfer. This vector must have names \code{"ERR"} and \code{"EAR"}, and weights must be within [0, 1] as well as sum to 1.
        \item \code{lat_t0}: Parameter for the S-shaped latency function - central value.
        \item \code{lat_eta}: Parameter for the S-shaped latency function - width of the transition period from 0 to full risk if argument \code{lat_method} is set to \code{"ProZES"}.
      }
    }
}
\references{
.
}
\seealso{
\code{\link{gen_exposure}},
\code{\link{sim_exposure}},
\code{\link{get_lear_indiv}},
\code{\link{get_lear_mc}},
\code{\link{get_lear_pop}}
}
\examples{
rm <- list(all_solid =rm_solid_incid_walsh2021(),
           breast    =rm_breast_incid_walsh2021(),
           leuk_lymph=rm_leuk_incid_walsh2021())

rm_mort <- list(all_solid =rm_solid_mort_sumray(),
                breast    =rm_breast_incid_walsh2021(),
                leuk_lymph=rm_leuk_incid_walsh2021())

base_cancer <- list(all_solid =d_cancer_ger_incid_solidW_i,
                    breast    =d_cancer_ger_incid_breastW_i,
                    leuk_lymph=d_cancer_ger_incid_leuk_lymphW_i)

base_cancer_mort <- list(all_solid =d_cancer_ger_mort_solidW_i,
                         breast    =d_cancer_ger_mort_breastW_i,
                         leuk_lymph=d_cancer_ger_mort_leuk_lymphW_i)

wt_transfer <- list(all_solid =c(ERR=0.5, EAR=0.5),
                    breast    =c(ERR=0.0, EAR=1.0),
                    leuk_lymph=c(ERR=1.0, EAR=0.0))

lat_t0 <- list(all_solid =4,
               breast    =4,
               leuk_lymph=1.5)

lat_eta <- list(all_solid =6.25,
                breast    =6.25,
                leuk_lymph=7.66)

expo_event <- gen_exposure(n          =2,
                           sex        ="f",
                           agex       =20,
                           dose_distr ="fixed",
                           dose_param =0.5,
                           ddref      =1,
                           dose_rate  ="acute",
                           cancer_site=list("breast",
                                            c("all_solid", "breast")))

p_mc <- gen_param_mc(exposure         =expo_event,
                     n_sim            =10L,
                     wt_transfer      =wt_transfer,
                     lat_t0           =lat_t0,
                     lat_eta          =lat_eta,
                     lat_method       ="ProZES",
                     lat_fixed        =FALSE,
                     wt_transfer_fixed=FALSE,
                     ddref_fixed      =FALSE,
                     risk_model       =rm,
                     risk_model_mort  =rm_mort)

p_mc[[1]]
}
